Refactor EnvironmentVariableRestore to testing library.
This class will be used in a future test related to #540, I think
a pure refactor PR is easier to grok.
