+++ google/cloud/CMakeLists.txt
@@ -90,10 +90,11 @@ include(CreateBazelConfig)
 create_bazel_config(google_cloud_cpp_common)
 google_cloud_cpp_add_clang_tidy(google_cloud_cpp_common)
 
-add_library(google_cloud_cpp_testing INTERFACE)
-target_sources(google_cloud_cpp_testing INTERFACE
-    cloud_testing/check_predicate_becomes_false.h)
-target_link_libraries(google_cloud_cpp_testing INTERFACE
+add_library(google_cloud_cpp_testing
+    cloud_testing/check_predicate_becomes_false.h
+    cloud_testing/environment_variable_restore.h
+    cloud_testing/environment_variable_restore.cc)
+target_link_libraries(google_cloud_cpp_testing PUBLIC
     google_cloud_cpp_common gmock)
 
 create_bazel_config(google_cloud_cpp_testing)
